requirements.txt
customtkinter>=5.2.0
sounddevice>=0.4.6
numpy>=1.24.0
faster-whisper>=1.0.0
pynput>=1.7.6
pyperclip>=1.8.2
requests>=2.31.0
platformdirs>=4.0.0
pystray>=0.19.0
Pillow>=10.0.0
